used uint16_t behavior to fix out-of-range phase in wave functions

+++ src/wave.c
@@ -17,7 +17,7 @@ void wave_init()
 /*  .-.
  --/---\---/--
         '-'  */
-float wave_sine(int off) {
+float wave_sine(uint16_t off) {
     div_t d = div(off, LUTSIZE);
     int index = (d.quot % 2 == 0) ? d.rem : LUTSIZE - d.rem - 1;
     int negative = (d.quot % 4 < 2) ? 1 : -1;
@@ -27,7 +27,7 @@ float wave_sine(int off) {
 /*  .-.
  --/---\___/--
              */
-float wave_halfsine(int off) {
+float wave_halfsine(uint16_t off) {
     div_t d = div(off, LUTSIZE);
     if(d.quot % 4 > 2) return 0;
     int index = (d.quot % 2 == 0) ? d.rem : LUTSIZE - d.rem - 1;
@@ -37,7 +37,7 @@ float wave_halfsine(int off) {
 /*  .-. .-.
  --/---V---\--
              */
-float wave_camelsine(int off) {
+float wave_camelsine(uint16_t off) {
     div_t d = div(off, LUTSIZE);
     int index = (d.quot % 2 == 0) ? d.rem : LUTSIZE - d.rem - 1;
     return lut[index];
@@ -46,7 +46,7 @@ float wave_camelsine(int off) {
 /*  .-  .-
  --/-|_/-|_/--
              */
-float wave_quartersine(int off) {
+float wave_quartersine(uint16_t off) {
     div_t d = div(off, LUTSIZE);
     return (d.quot % 2 == 0) ? lut[d.rem] : 0;
 }
@@ -54,13 +54,13 @@ float wave_quartersine(int off) {
 /* .---.
  --|---|---|--
        '---' */
-float wave_pulse(int off) {
+float wave_pulse(uint16_t off) {
     return off < INT16_MAX/4 ? 1 : -1;
 }
 
 /*  /|  /|
  --/-|-/-|-/--
      |/  |/ */
-float wave_ramp(int off) {
+float wave_ramp(uint16_t off) {
     return fmodf(off / (float)INT16_MAX, 1.0);
 }
